requirements.txt
appdirs==1.4.4
asgiref==3.5.2
CacheControl==0.12.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
Cheetah3==3.2.6.post2
click==8.0.3
colorama==0.4.4
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
distlib==0.3.4
distro==1.7.0
Django==4.0.4
djangorestframework==3.13.1
dnspython==2.2.1
docopt==0.6.2
drf-yasg==1.20.0
Flask==2.0.2
greenlet==1.1.2
html5lib==1.1
idna==3.3
importlib-metadata==4.8.1
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.1.1
msgpack==1.0.4
packaging==21.3
psycopg2==2.9.3
pycparser==2.21
pyparsing==3.0.9
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
webencodings==0.5.1
Werkzeug==2.1.2
zipp==3.8.0
gunicorn==20.0.4
